requirements.txt
dash
opencv-python
dlib
deepface
numpy
Pillow
gdown
